backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
browser-use==1.0.0
langchain-google-genai==1.0.8
pydantic==2.5.0
asyncio
threading
base64